Move selectionRequired to shared
